solved ini4
Also, I added a check to add one to the first number if it is even. This way, all of the even numbers get skipped in the for loop. I know it does not really make a speed difference, but I wanted to think about how I could make it faster.